requirements.txt
mcp
fastapi
uvicorn[standard]
requests
elasticsearch
python-dotenv
faker